src/data_retrieval/nasa_api.py
"""
NASA API Data Retriever Module.
Handles data retrieval from NASA Earth Data APIs.
"""

import os
import logging
from datetime import datetime, timedelta
from typing import Dict, List, Any, Optional
import numpy as np
import pandas as pd
import requests
from dotenv import load_dotenv

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class NASADataRetriever:
    """
    Handles data retrieval from NASA Earth Data APIs.
    Provides methods for fetching environmental data.
    """
    
    def __init__(self, config: Dict = None):
        self.config = config or {}
        
        # Load environment variables
        load_dotenv()
        
        # Set up API credentials
        self.api_key = os.getenv('NASA_API_KEY', 'DEMO_KEY')
        self.base_url = 'https://cmr.earthdata.nasa.gov/search'
        
        # Initialize session
        self.session = requests.Session()
        if 'username' in self.config and 'password' in self.config:
            self.session.auth = (
                self.config['username'],
                self.config['password']
            )
        
        logger.info("NASA Data Retriever initialized")
        logger.info(f"Using API base URL: {self.base_url}")
        
        # Test connection
        try:
            self.session.get(self.base_url)
            logger.info("Successfully connected to NASA Earthdata API")
        except requests.RequestException as e:
            logger.warning(f"Could not connect to NASA Earthdata API: {e}")
    
    def get_data(
        self,
        lat: float,
        lon: float,
        radius: float,
        start_date: datetime,
        end_date: datetime
    ) -> Dict[str, np.ndarray]:
        """
        Retrieve environmental data for a location.
        
        Args:
            lat: Latitude
            lon: Longitude
            radius: Search radius in km
            start_date: Start date
            end_date: End date
            
        Returns:
            Dictionary of environmental data arrays
        """
        logger.info(
            f"Retrieving data for 4 parameters from "
            f"{start_date.strftime('%Y-%m-%d')} to "
            f"{end_date.strftime('%Y-%m-%d')}"
        )
        
        # Initialize data dictionary
        data = {}
        
        # Get LST data
        logger.info("Generating synthetic data for LST")
        logger.warning("Generating synthetic data for LST")
        data['LST'] = self._generate_synthetic_data(
            start_date,
            end_date,
            base_value=298,  # ~25°C
            amplitude=10,
            noise_level=2
        )
        
        # Get NDVI data
        logger.info("Generating synthetic data for NDVI")
        logger.warning("Generating synthetic data for NDVI")
        data['NDVI'] = self._generate_synthetic_data(
            start_date,
            end_date,
            base_value=0.5,
            amplitude=0.2,
            noise_level=0.1
        )
        
        # Get precipitation data
        logger.info("Generating synthetic data for Precipitation")
        logger.warning("Generating synthetic data for Precipitation")
        data['Precipitation'] = self._generate_synthetic_data(
            start_date,
            end_date,
            base_value=2,
            amplitude=5,
            noise_level=1,
            allow_negative=False
        )
        
        # Get soil moisture data
        logger.info("Generating synthetic data for SoilMoisture")
        logger.warning("Generating synthetic data for SoilMoisture")
        data['SoilMoisture'] = self._generate_synthetic_data(
            start_date,
            end_date,
            base_value=0.3,
            amplitude=0.1,
            noise_level=0.05,
            allow_negative=False
        )
        
        return data
    
    def _generate_synthetic_data(
        self,
        start_date: datetime,
        end_date: datetime,
        base_value: float = 0,
        amplitude: float = 1,
        noise_level: float = 0.1,
        allow_negative: bool = True
    ) -> np.ndarray:
        """
        Generate synthetic data for testing.
        
        Args:
            start_date: Start date
            end_date: End date
            base_value: Base value for the data
            amplitude: Amplitude of seasonal variation
            noise_level: Level of random noise
            allow_negative: Whether to allow negative values
            
        Returns:
            Array of synthetic data
        """
        # Generate time points
        days = (end_date - start_date).days
        t = np.linspace(0, 2 * np.pi, days)
        
        # Generate seasonal pattern
        seasonal = amplitude * np.sin(t)
        
        # Add random noise
        noise = np.random.normal(0, noise_level, days)
        
        # Combine components
        data = base_value + seasonal + noise
        
        # Apply constraints
        if not allow_negative:
            data = np.maximum(data, 0)
        
        return data
    
    def _fetch_nasa_data(
        self,
        dataset: str,
        lat: float,
        lon: float,
        radius: float,
        start_date: str,
        end_date: str
    ) -> Optional[Dict[str, Any]]:
        """
        Fetch data from NASA Earth Data API.
        
        Args:
            dataset: Dataset identifier
            lat: Latitude
            lon: Longitude
            radius: Search radius in km
            start_date: Start date string (YYYY-MM-DD)
            end_date: End date string (YYYY-MM-DD)
            
        Returns:
            API response data or None if request fails
        """
        try:
            params = {
                'dataset': dataset,
                'lat': lat,
                'lon': lon,
                'radius': radius,
                'start_date': start_date,
                'end_date': end_date,
                'api_key': self.api_key
            }
            
            response = self.session.get(
                f"{self.base_url}/granules",
                params=params
            )
            response.raise_for_status()
            
            return response.json()
        
        except requests.RequestException as e:
            logger.error(f"Error fetching NASA data: {e}")
            return None
    
    def _process_nasa_response(
        self,
        response_data: Dict[str, Any]
    ) -> Optional[np.ndarray]:
        """
        Process NASA API response data.
        
        Args:
            response_data: API response data
            
        Returns:
            Processed data array or None if processing fails
        """
        try:
            # Extract and process data
            # This is a placeholder - implement actual processing
            # based on the specific dataset format
            return np.array([])
            
        except Exception as e:
            logger.error(f"Error processing NASA data: {e}")
            return None 