requirements.txt
# Core Data Processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=0.24.2
xarray>=0.19.0

# Machine Learning
torch>=1.9.0
transformers>=4.11.0
sentence-transformers>=2.1.0

# Data Visualization
matplotlib>=3.4.3
seaborn>=0.11.2
plotly>=5.3.1
folium>=0.12.1

# Web Framework
flask>=2.0.1
flask-cors>=3.0.10
gunicorn>=20.1.0

# NASA Earth Data API
earthengine-api>=0.1.290
netCDF4>=1.5.7
requests>=2.26.0

# Geospatial Processing
rasterio>=1.2.10
geopandas>=0.9.0
shapely>=1.8.0

# Utilities
pyyaml>=5.4.1
python-dotenv>=0.19.0
tqdm>=4.62.3
joblib>=1.0.1

# Testing and Development
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2
mypy>=0.910 